api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.1
supabase>=2.3.0
python-dotenv>=1.0.0
httpx>=0.26.0
pydantic>=2.5.0
